+++ tests/SnapshotTests/EfCoreGeneration/snapshots/snap-v8.0/EfCoreGenerationTests.Writes_efcore_converters_that_respect_namespaces.verified.txt
